semana03/dia3/requirements.txt
#framework principal
flask

#extensiones de flask
flask-mysqldb

#otras librerias